+++ component/pom.xml
@@ -20,9 +20,9 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.extension.siddhi.io.sqs</groupId>
+        <groupId>io.siddhi.extension.io.sqs</groupId>
         <artifactId>siddhi-io-sqs-parent</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>2.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>siddhi-io-sqs</artifactId>
@@ -54,12 +54,12 @@
             <artifactId>testng</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+            <groupId>io.siddhi.extension.map.xml</groupId>
             <artifactId>siddhi-map-xml</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.map.json</groupId>
+            <groupId>io.siddhi.extension.map.json</groupId>
             <artifactId>siddhi-map-json</artifactId>
             <scope>test</scope>
         </dependency>
@@ -224,7 +224,7 @@
                             com.fasterxml.jackson.annotation.*
                         </Private-Package>
                         <Export-Package>
-                            org.wso2.extension.siddhi.io.sqs.*;version="${project.version}",
+                            io.siddhi.extension.io.sqs.*;version="${project.version}",
                         </Export-Package>
                         <Import-Package>
                             com.amazonaws.*;version="${aws.java.sqs.sdk.version.range}",
