--- docs/stylesheets/extra.css
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-.md-header-nav__button.md-logo img {
-    width: 140px;
-    height: 33px;
-    margin-right: 0;
-}
-
-.md-content__icon,
-.md-footer-nav__button,
-.md-header-nav__button,
-.md-nav__button,
-.md-nav__title::before,
-.md-search-result__article--document::before {
-    margin: 0.3rem;
-    padding: 0;
-}
-
-.extention-title {
-    font-weight: 700;
-    margin-right: 50px;
-}
-
-.md-header-nav__title {
-    padding-left: 5px;
-}
-
-@media (max-width: 1219px) {
-
-    .extention-title {
-        display: none;
-    }
-
-    html .md-nav--primary .md-nav__title--site .md-nav__button {
-        width: 13.4rem;
-    }
-}
