+++ mkdocs.yml
@@ -1,36 +1,37 @@
 site_name: Siddhi IO SQS
-site_description: Siddhi IO sqs Extension
-site_author: WSO2
-site_url: https://wso2-extensions.github.io/siddhi-io-sqs/
-extra_css:
-- stylesheets/extra.css
-extra_javascript:
-- javascripts/extra.js
+site_description: Siddhi IO SQS Extension
 repo_name: Siddhi-IO-SQS
-repo_url: https://github.com/wso2-extensions/siddhi-io-sqs
-copyright: Copyright &copy; 2011 - 2018 WSO2
+repo_url: https://github.com/siddhi.io/siddhi-io-sqs
+edit_uri: https://github.com/siddhi.io/siddhi-io-sqs/blob/master
+copyright: Siddhi - Documentation
 theme:
   name: material
-  logo: images/siddhi-logo-w.svg
+  logo: images/siddhi-logo.svg
   favicon: images/favicon.ico
   palette:
     primary: teal
     accent: teal
-google_analytics:
-- UA-103065-28
-- auto
+extra_css:
+  - assets/stylesheets/extra.css
+extra_javascript:
+  - assets/javascripts/extra.js
 extra:
   social:
-  - type: github
-    link: https://github.com/wso2/siddhi
-  - type: linkedin
-    link: https://www.linkedin.com/groups/13553064
+    - type: github
+      link: https://github.com/siddhi-io/siddhi
+    - type: medium
+      link: https://medium.com/siddhi-io
+    - type: twitter
+      link: https://twitter.com/siddhi_io
+    - type: linkedin
+      link: https://www.linkedin.com/groups/13553064
+google_analytics:
+  - UA-103065-28
+  - auto
 markdown_extensions:
-- admonition
-- toc(permalink=true)
-- codehilite(guess_lang=false)
+  - markdown.extensions.admonition
 pages:
-- Welcome: index.md
+- Information: index.md
 - API Docs:
   - 2.0.0: api/2.0.0.md
   - 1.1.0: api/1.1.0.md
