+++ pom.xml
@@ -29,9 +29,9 @@
     <modules>
         <module>component</module>
     </modules>
-    <groupId>org.wso2.extension.siddhi.io.sqs</groupId>
+    <groupId>io.siddhi.extension.io.sqs</groupId>
     <artifactId>siddhi-io-sqs-parent</artifactId>
-    <version>2.0.1-SNAPSHOT</version>
+    <version>2.1.0-SNAPSHOT</version>
     <name>Siddhi I/O sqs</name>
     <dependencyManagement>
         <dependencies>
@@ -79,13 +79,13 @@
                 </exclusions>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+                <groupId>io.siddhi.extension.map.xml</groupId>
                 <artifactId>siddhi-map-xml</artifactId>
                 <version>${siddhi.map.xml.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.json</groupId>
+                <groupId>io.siddhi.extension.map.json</groupId>
                 <artifactId>siddhi-map-json</artifactId>
                 <version>${siddhi.map.json.version}</version>
                 <scope>test</scope>
@@ -210,11 +210,11 @@
     </scm>
 
     <properties>
-        <siddhi.version>5.0.0</siddhi.version>
+        <siddhi.version>5.1.2</siddhi.version>
         <log4j.version>1.2.17.wso2v1</log4j.version>
         <carbon.transport.version>4.4.15</carbon.transport.version>
-        <siddhi.map.json.version>4.0.31</siddhi.map.json.version>
-        <siddhi.map.xml.version>4.0.18</siddhi.map.xml.version>
+        <siddhi.map.json.version>5.0.4</siddhi.map.json.version>
+        <siddhi.map.xml.version>5.0.3</siddhi.map.xml.version>
         <jacoco.plugin.version>0.7.9</jacoco.plugin.version>
         <carbon.feature.plugin.version>3.0.0</carbon.feature.plugin.version>
         <aws.java.sqs.sdk.version>1.11.285</aws.java.sqs.sdk.version>
